Prepare outPt class to data locality
